+++ src/Pim/Bundle/UserBundle/Controller/Rest/UserGroupController.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\UserBundle\Controller\Rest;
 
-use Pim\Bundle\UserBundle\Doctrine\ORM\Repository\GroupRepository;
+use Pim\Bundle\UserBundle\Persistence\ORM\Repository\GroupRepository;
 use Pim\Component\User\Group\GroupInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
