+++ src/Pim/Bundle/UserBundle/Resources/config/model/doctrine/Role.orm.yml
@@ -2,7 +2,7 @@ Pim\Component\User\Model\Role:
     type: entity
     table: oro_access_role
     changeTrackingPolicy: DEFERRED_EXPLICIT
-    repositoryClass: Pim\Bundle\UserBundle\Doctrine\ORM\Repository\RoleRepository
+    repositoryClass: Pim\Bundle\UserBundle\Persistence\ORM\Repository\RoleRepository
     fields:
         id:
             type: integer
