+++ src/Pim/Bundle/UserBundle/Resources/config/model/doctrine/User.orm.yml
@@ -2,7 +2,7 @@ Pim\Component\User\Model\User:
     type: entity
     table: oro_user
     changeTrackingPolicy: DEFERRED_EXPLICIT
-    repositoryClass: Pim\Bundle\UserBundle\Doctrine\ORM\Repository\UserRepository
+    repositoryClass: Pim\Bundle\UserBundle\Persistence\ORM\Repository\UserRepository
     fields:
         id:
             type: integer
